+++ FO-DICOM.Core/Network/Client/DicomClient.cs
@@ -5,6 +5,7 @@ using FellowOakDicom.Network.Client.Advanced.Association;
 using FellowOakDicom.Network.Client.Advanced.Connection;
 using FellowOakDicom.Network.Client.EventArguments;
 using FellowOakDicom.Network.Client.States;
+using FellowOakDicom.Network.Tls;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
@@ -35,9 +36,9 @@ namespace FellowOakDicom.Network.Client
         int Port { get; }
 
         /// <summary>
-        /// True if TLS security should be enabled, false otherwise.
+        /// A handler to initiate TLS security, if null then TLS is not enabled.
         /// </summary>
-        bool UseTls { get; }
+        ITlsInitiator TlsInitiator { get; }
 
         /// <summary>
         /// Calling Application Entity Title.
@@ -172,7 +173,7 @@ namespace FellowOakDicom.Network.Client
 
         public string Host { get; }
         public int Port { get; }
-        public bool UseTls { get; }
+        public ITlsInitiator TlsInitiator { get; }
         public string CallingAe { get; }
         public string CalledAe { get; }
         
@@ -205,14 +206,14 @@ namespace FellowOakDicom.Network.Client
         /// </summary>
         /// <param name="host">DICOM host.</param>
         /// <param name="port">Port.</param>
-        /// <param name="useTls">True if TLS security should be enabled, false otherwise.</param>
+        /// <param name="tlsInitiator">if null then no TLS is enabled, otherwise the handler to initiate TLS security.</param>
         /// <param name="callingAe">Calling Application Entity Title.</param>
         /// <param name="calledAe">Called Application Entity Title.</param>
         /// <param name="clientOptions">The options that further modify the behavior of this DICOM client</param>
         /// <param name="serviceOptions">The options that modify the behavior of the base DICOM service</param>
         /// <param name="loggerFactory">The log manager that will be used to extract a default logger</param>
         /// <param name="advancedDicomClientConnectionFactory">The advanced DICOM client factory that will be used to actually send the requests</param>
-        public DicomClient(string host, int port, bool useTls, string callingAe, string calledAe,
+        public DicomClient(string host, int port, ITlsInitiator tlsInitiator, string callingAe, string calledAe,
             DicomClientOptions clientOptions,
             DicomServiceOptions serviceOptions,
             ILoggerFactory loggerFactory,
@@ -220,7 +221,7 @@ namespace FellowOakDicom.Network.Client
         {
             Host = host;
             Port = port;
-            UseTls = useTls;
+            TlsInitiator = tlsInitiator;
             CallingAe = callingAe;
             CalledAe = calledAe;
             ClientOptions = clientOptions;
@@ -311,9 +312,8 @@ namespace FellowOakDicom.Network.Client
                             {
                                 Host = Host,
                                 Port = Port,
-                                UseTls = UseTls,
+                                TlsInitiator = TlsInitiator,
                                 NoDelay = ServiceOptions.TcpNoDelay,
-                                IgnoreSslPolicyErrors = ServiceOptions.IgnoreSslPolicyErrors,
                                 Timeout = TimeSpan.FromMilliseconds(ClientOptions.AssociationRequestTimeoutInMs)
                             },
                             RequestHandlers = new AdvancedDicomClientConnectionRequestHandlers