+++ FO-DICOM.Core/Network/DicomServiceOptions.cs
@@ -28,9 +28,6 @@ namespace FellowOakDicom.Network
         /// <summary>Gets or sets maximum buffer length for data PDVs when generating P-Data-TF PDUs.</summary>
         public uint MaxDataBuffer { get; set; } = 1 * 1024 * 1024; //1MB
 
-        /// <summary>Gets or sets whether DICOM client should ignore SSL certificate errors.</summary>
-        public bool IgnoreSslPolicyErrors { get; set; } = false;
-
         /// <summary>Gets or sets whether to enable (true) or disable (false) TCP Nagle algorithm.</summary>
         public bool TcpNoDelay { get; set; } = true;
 
@@ -69,7 +66,6 @@ namespace FellowOakDicom.Network
                 UseRemoteAEForLogName = UseRemoteAEForLogName,
                 MaxCommandBuffer = MaxCommandBuffer,
                 MaxDataBuffer = MaxDataBuffer,
-                IgnoreSslPolicyErrors = IgnoreSslPolicyErrors,
                 TcpNoDelay = TcpNoDelay,
                 RequestTimeout = RequestTimeout,
                 MaxClientsAllowed = MaxClientsAllowed,